extract RUM logic out of Selenium contrib, reenable testing selenium/capybara integration for Ruby 3.4
