+++ lib/datadog/ci/contrib/selenium/configuration/settings.rb
@@ -5,6 +5,8 @@ require "datadog/core"
 require_relative "../ext"
 require_relative "../../settings"
 
+require_relative "../../../ext/rum"
+
 module Datadog
   module CI
     module Contrib
@@ -21,7 +23,7 @@ module Datadog
 
             option :rum_flush_wait_millis do |o|
               o.type :int
-              o.env Ext::ENV_RUM_FLUSH_WAIT_MILLIS
+              o.env CI::Ext::RUM::ENV_RUM_FLUSH_WAIT_MILLIS
               o.default 500
             end
           end
