+++ lib/datadog/ci/contrib/selenium/driver.rb
@@ -2,8 +2,8 @@
 
 require_relative "../patcher"
 
-require_relative "ext"
-require_relative "rum"
+require_relative "../../utils/rum"
+require_relative "../../ext/rum"
 require_relative "../../ext/test"
 
 module Datadog
@@ -22,10 +22,10 @@ module Datadog
 
               Datadog.logger.debug("[Selenium] Driver quit event")
 
-              RUM.stop_rum_session(@bridge)
+              Utils::RUM.stop_rum_session(@bridge)
 
               Datadog.logger.debug("[Selenium] RUM session stopped, deleting cookie")
-              @bridge.manage.delete_cookie(Ext::COOKIE_TEST_EXECUTION_ID)
+              @bridge.manage.delete_cookie(CI::Ext::RUM::COOKIE_TEST_EXECUTION_ID)
             rescue ::Selenium::WebDriver::Error::WebDriverError => e
               Datadog.logger.debug("[Selenium] Error while quitting Selenium driver: #{e.message}")
             ensure
