+++ sig/datadog/ci/contrib/selenium/ext.rbs
@@ -4,14 +4,6 @@ module Datadog
       module Selenium
         module Ext
           ENV_ENABLED: "DD_CIVISIBILITY_SELENIUM_ENABLED"
-
-          ENV_RUM_FLUSH_WAIT_MILLIS: "DD_CIVISIBILITY_RUM_FLUSH_WAIT_MILLIS"
-
-          COOKIE_TEST_EXECUTION_ID: "datadog-ci-visibility-test-execution-id"
-
-          SCRIPT_IS_RUM_ACTIVE: String
-
-          SCRIPT_STOP_RUM_SESSION: String
         end
       end
     end
