--- sig/datadog/ci/contrib/selenium/rum.rbs
@@ -1,13 +1,0 @@
-module Datadog
-  module CI
-    module Contrib
-      module Selenium
-        module RUM
-          def self.is_rum_active?: (untyped script_executor) -> bool
-
-          def self.stop_rum_session: (untyped script_executor) -> void
-        end
-      end
-    end
-  end
-end
