+++ sig/datadog/ci/ext/rum.rbs
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+module Datadog
+  module CI
+    module Ext
+      module RUM
+        ENV_RUM_FLUSH_WAIT_MILLIS: "DD_CIVISIBILITY_RUM_FLUSH_WAIT_MILLIS"
+
+        COOKIE_TEST_EXECUTION_ID: "datadog-ci-visibility-test-execution-id"
+
+        SCRIPT_IS_RUM_ACTIVE: String
+        SCRIPT_STOP_RUM_SESSION: String
+      end
+    end
+  end
+end
