+++ spec/datadog/ci/contrib/selenium/instrumentation_spec.rb
@@ -78,12 +78,12 @@ RSpec.describe "Browser tests with selenium" do
     expect(visited_urls).to eq(["http://www.example.com", "about:blank"])
     expect(executed_scripts).to eq(
       [
-        Datadog::CI::Contrib::Selenium::Ext::SCRIPT_IS_RUM_ACTIVE,
-        Datadog::CI::Contrib::Selenium::Ext::SCRIPT_STOP_RUM_SESSION,
+        Datadog::CI::Ext::RUM::SCRIPT_IS_RUM_ACTIVE,
+        Datadog::CI::Ext::RUM::SCRIPT_STOP_RUM_SESSION,
         "window.sessionStorage.clear()",
         "window.localStorage.clear()",
-        Datadog::CI::Contrib::Selenium::Ext::SCRIPT_IS_RUM_ACTIVE,
-        Datadog::CI::Contrib::Selenium::Ext::SCRIPT_STOP_RUM_SESSION
+        Datadog::CI::Ext::RUM::SCRIPT_IS_RUM_ACTIVE,
+        Datadog::CI::Ext::RUM::SCRIPT_STOP_RUM_SESSION
       ]
     )
